+++ ngrok/src/session.rs
@@ -1,6 +1,7 @@
 use std::{
     collections::HashMap,
     env,
+    future::Future,
     io,
     num::ParseIntError,
     sync::Arc,
@@ -12,12 +13,14 @@ use async_rustls::rustls::{
     self,
     client::InvalidDnsNameError,
 };
+use async_trait::async_trait;
 use futures::{
     future,
-    future::BoxFuture,
+    prelude::*,
     FutureExt,
 };
 use muxado::heartbeat::HeartbeatConfig;
+pub use muxado::heartbeat::HeartbeatHandler;
 use rustls_pemfile::Item;
 use thiserror::Error;
 use tokio::{
@@ -36,7 +39,7 @@ use tokio::{
 };
 use tokio_retry::{
     strategy::ExponentialBackoff,
-    Retry,
+    RetryIf,
 };
 use tokio_util::compat::{
     FuturesAsyncReadCompatExt,
@@ -132,43 +135,72 @@ struct SessionInner {
 pub trait IoStream: AsyncRead + AsyncWrite + Unpin + Send + 'static {}
 impl<T> IoStream for T where T: AsyncRead + AsyncWrite + Unpin + Send + 'static {}
 
-/// The function used to establish the connection to the ngrok server.
+/// Trait for establishing the connection to the ngrok server.
+#[async_trait]
+pub trait Connector: Sync + Send + 'static {
+    /// The function used to establish the connection to the ngrok server.
+    ///
+    /// This is effectively `async |addr, tls_config, err| -> Result<IoStream>`.
+    ///
+    /// If it is being called due to a disconnect, the [AcceptError] argument will
+    /// be populated.
+    ///
+    /// If it returns `Err(ConnectError::Canceled)`, reconnecting will be canceled
+    /// and the session will be terminated. Note that this error will never be
+    /// returned from the [default_connect] function.
+    async fn connect(
+        &self,
+        addr: String,
+        tls_config: Arc<rustls::ClientConfig>,
+        err: Option<AcceptError>,
+    ) -> Result<Box<dyn IoStream>, ConnectError>;
+}
+
+#[async_trait]
+impl<F, U> Connector for F
+where
+    F: Fn(String, Arc<rustls::ClientConfig>, Option<AcceptError>) -> U + Send + Sync + 'static,
+    U: Future<Output = Result<Box<dyn IoStream>, ConnectError>> + Send,
+{
+    async fn connect(
+        &self,
+        addr: String,
+        tls_config: Arc<rustls::ClientConfig>,
+        err: Option<AcceptError>,
+    ) -> Result<Box<dyn IoStream>, ConnectError> {
+        self(addr, tls_config, err).await
+    }
+}
+
+/// The default ngrok connector.
 ///
-/// This is effectively `async |addr, tls_config| -> Result<IoStream>`, with all
-/// of the necessary boxing to turn the generics into trait objects.
-pub type ConnectFn = Arc<
-    dyn Fn(
-            String,
-            Arc<rustls::ClientConfig>,
-        ) -> BoxFuture<'static, Result<Box<dyn IoStream>, ConnectError>>
-        + Send
-        + Sync
-        + 'static,
->;
+/// Establishes a TCP connection to `addr`, and then performs a TLS handshake
+/// using the `tls_config`.
+///
+/// Discards any errors during reconnect, allowing attempts to recur
+/// indefinitely.
+pub async fn default_connect(
+    addr: String,
+    tls_config: Arc<rustls::ClientConfig>,
+    _: Option<AcceptError>,
+) -> Result<Box<dyn IoStream>, ConnectError> {
+    let mut split = addr.split(':');
+    let host = split.next().unwrap();
+    let port = split
+        .next()
+        .map(str::parse::<u16>)
+        .transpose()?
+        .unwrap_or(443);
+    let conn = tokio::net::TcpStream::connect(&(host, port))
+        .await
+        .map_err(ConnectError::Tcp)?
+        .compat();
 
-fn default_connect() -> ConnectFn {
-    Arc::new(|addr, tls_config| {
-        async move {
-            let mut split = addr.split(':');
-            let host = split.next().unwrap();
-            let port = split
-                .next()
-                .map(str::parse::<u16>)
-                .transpose()?
-                .unwrap_or(443);
-            let conn = tokio::net::TcpStream::connect(&(host, port))
-                .await
-                .map_err(ConnectError::Tcp)?
-                .compat();
-
-            let tls_conn = async_rustls::TlsConnector::from(tls_config)
-                .connect(rustls::ServerName::try_from(host)?, conn)
-                .await
-                .map_err(ConnectError::Tls)?;
-            Ok(Box::new(tls_conn.compat()) as Box<dyn IoStream>)
-        }
-        .boxed()
-    })
+    let tls_conn = async_rustls::TlsConnector::from(tls_config)
+        .connect(rustls::ServerName::try_from(host)?, conn)
+        .await
+        .map_err(ConnectError::Tls)?;
+    Ok(Box::new(tls_conn.compat()) as Box<dyn IoStream>)
 }
 
 /// The builder for an ngrok [Session].
@@ -178,9 +210,10 @@ pub struct SessionBuilder {
     metadata: Option<String>,
     heartbeat_interval: Option<Duration>,
     heartbeat_tolerance: Option<Duration>,
+    heartbeat_handler: Option<Arc<dyn HeartbeatHandler>>,
     server_addr: String,
     tls_config: rustls::ClientConfig,
-    connector: ConnectFn,
+    connector: Arc<dyn Connector>,
     handlers: CommandHandlers,
     cookie: Option<SecretString>,
     id: Option<String>,
@@ -226,6 +259,15 @@ pub enum ConnectError {
     /// An error occurred when attempting to authenticate.
     #[error("authentication failure")]
     Auth(RpcError),
+    /// An error occurred when rebinding tunnels during a reconnect
+    #[error("error rebinding tunnel after reconnect")]
+    Rebind(RpcError),
+    /// The (re)connect function gave up.
+    ///
+    /// This will never be returned by the default connect function, and is
+    /// instead used to cancel the reconnect loop.
+    #[error("the connect function gave up")]
+    Canceled,
 }
 
 impl Default for SessionBuilder {
@@ -254,9 +296,10 @@ impl Default for SessionBuilder {
             metadata: None,
             heartbeat_interval: None,
             heartbeat_tolerance: None,
+            heartbeat_handler: None,
             server_addr: "tunnel.ngrok.com:443".into(),
             tls_config,
-            connector: default_connect(),
+            connector: Arc::new(default_connect),
             handlers: Default::default(),
             cookie: None,
             id: None,
@@ -328,7 +371,7 @@ impl SessionBuilder {
     /// See the [server_addr parameter in the ngrok docs] for additional details.
     ///
     /// [server_addr parameter in the ngrok docs]: https://ngrok.com/docs/ngrok-agent/config#server_addr
-    pub fn server_addr(&mut self, addr: impl Into<String>) -> &mut Self {
+    pub fn server_addr(mut self, addr: impl Into<String>) -> Self {
         self.server_addr = addr.into();
         self
     }
@@ -351,8 +394,8 @@ impl SessionBuilder {
     /// ngrok service. Use this option if you need to connect through an outbound
     /// proxy. In the event of network disruptions, it will be called each time
     /// the session reconnects.
-    pub fn connector(&mut self, connect: ConnectFn) -> &mut Self {
-        self.connector = connect;
+    pub fn connector(mut self, connect: impl Connector) -> Self {
+        self.connector = Arc::new(connect);
         self
     }
 
@@ -366,7 +409,7 @@ impl SessionBuilder {
     /// Do not block inside this callback. It will cause the Dashboard or API
     /// stop operation to time out. Do not call [std::process::exit] inside this
     /// callback, it will also cause the operation to time out.
-    pub fn handle_stop_command(&mut self, handler: impl CommandHandler<Stop>) -> &mut Self {
+    pub fn handle_stop_command(mut self, handler: impl CommandHandler<Stop>) -> Self {
         self.handlers.on_stop = Some(Arc::new(handler));
         self
     }
@@ -382,7 +425,7 @@ impl SessionBuilder {
     /// Do not block inside this callback. It will cause the Dashboard or API
     /// stop operation to time out. Do not call [std::process::exit] inside this
     /// callback, it will also cause the operation to time out.
-    pub fn handle_restart_command(&mut self, handler: impl CommandHandler<Restart>) -> &mut Self {
+    pub fn handle_restart_command(mut self, handler: impl CommandHandler<Restart>) -> Self {
         self.handlers.on_restart = Some(Arc::new(handler));
         self
     }
@@ -398,8 +441,17 @@ impl SessionBuilder {
     /// Do not block inside this callback. It will cause the Dashboard or API
     /// stop operation to time out. Do not call [std::process::exit] inside this
     /// callback, it will also cause the operation to time out.
-    pub fn handle_update_command(&mut self, handler: impl CommandHandler<Update>) -> &mut Self {
+    pub fn handle_update_command(mut self, handler: impl CommandHandler<Update>) -> Self {
         self.handlers.on_update = Some(Arc::new(handler));
+        self
+    }
+
+    /// Call the provided handler whenever a heartbeat response is received.
+    ///
+    /// If the handler returns an error, the heartbeat task will exit, resulting
+    /// in the session eventually dying as well.
+    pub fn handle_heartbeat(mut self, callback: impl HeartbeatHandler) -> Self {
+        self.heartbeat_handler = Some(Arc::new(callback));
         self
     }
 
@@ -408,7 +460,7 @@ impl SessionBuilder {
     /// an error.
     pub async fn connect(&self) -> Result<Session, ConnectError> {
         let (dropref, dropped) = awaitdrop::awaitdrop();
-        let (inner, incoming) = self.connect_inner().await?;
+        let (inner, incoming) = self.connect_inner(None).await?;
 
         let inner = Arc::new(ArcSwap::new(inner.into()));
 
@@ -420,17 +472,27 @@ impl SessionBuilder {
         Ok(Session { dropref, inner })
     }
 
-    async fn connect_inner(&self) -> Result<(SessionInner, IncomingStreams), ConnectError> {
-        let conn =
-            (self.connector)(self.server_addr.clone(), Arc::new(self.tls_config.clone())).await?;
+    async fn connect_inner(
+        &self,
+        err: impl Into<Option<AcceptError>>,
+    ) -> Result<(SessionInner, IncomingStreams), ConnectError> {
+        let conn = self
+            .connector
+            .connect(
+                self.server_addr.clone(),
+                Arc::new(self.tls_config.clone()),
+                err.into(),
+            )
+            .await?;
 
-        let mut heartbeat_config = HeartbeatConfig::<fn(Duration)>::default();
+        let mut heartbeat_config = HeartbeatConfig::default();
         if let Some(interval) = self.heartbeat_interval {
             heartbeat_config.interval = interval;
         }
         if let Some(tolerance) = self.heartbeat_tolerance {
             heartbeat_config.tolerance = tolerance;
         }
+        heartbeat_config.handler = self.heartbeat_handler.clone();
         // convert these while we have ownership
         let interval_nanos = heartbeat_config.interval.as_nanos();
         let heartbeat_interval = i64::try_from(interval_nanos)
@@ -671,13 +733,12 @@ async fn accept_one(
     Ok(())
 }
 
-async fn try_reconnect(inner: Arc<ArcSwap<SessionInner>>) -> Result<IncomingStreams, AcceptError> {
+async fn try_reconnect(
+    inner: Arc<ArcSwap<SessionInner>>,
+    err: impl Into<Option<AcceptError>>,
+) -> Result<IncomingStreams, ConnectError> {
     let old_inner = inner.load();
-    let (new_inner, new_incoming) = old_inner
-        .builder
-        .connect_inner()
-        .await
-        .map_err(|_| AcceptError::Transport(muxado::Error::SessionClosed))?;
+    let (new_inner, new_incoming) = old_inner.builder.connect_inner(err).await?;
     let mut client = new_inner.client.lock().await;
     let mut new_tunnels = new_inner.tunnels.write().await;
     let old_tunnels = old_inner.tunnels.read().await;
@@ -693,14 +754,14 @@ async fn try_reconnect(inner: Arc<ArcSwap<SessionInner>>) -> Result<IncomingStre
                     &tun.forwards_to,
                 )
                 .await
-                .map_err(|_| AcceptError::Transport(muxado::Error::ErrorUnknown))?;
+                .map_err(ConnectError::Rebind)?;
             debug!(?resp, %id, %tun.proto, ?tun.opts, ?tun.extra, %tun.forwards_to, "rebound tunnel");
             new_tunnels.insert(id.clone(), tun.clone());
         } else {
             let resp = client
                 .listen_label(tun.labels.clone(), &tun.extra.metadata, &tun.forwards_to)
                 .await
-                .map_err(|_| AcceptError::Transport(muxado::Error::ErrorUnknown))?;
+                .map_err(ConnectError::Rebind)?;
 
             if !resp.id.is_empty() {
                 new_tunnels.insert(resp.id, tun.clone());
@@ -722,19 +783,36 @@ async fn accept_incoming(mut incoming: IncomingStreams, inner: Arc<ArcSwap<Sessi
     let error: AcceptError = loop {
         if let Err(error) = accept_one(&mut incoming, &inner).await {
             debug!(%error, "failed to accept stream, attempting reconnect");
-            let reconnect = Retry::spawn(ExponentialBackoff::from_millis(50), || {
-                try_reconnect(inner.clone())
-            });
+            // This is gross, but should perform fine. Couple of notes:
+            // * Mutex so that both the action and condition can share access to
+            //   `error`. Realistically, the lock calls should be non-concurrent,
+            //   but Rust can't prove that.
+            // * Not setting the error in the action because then a a reference
+            //   to a FnMut closure would escape via the returned Future, which is
+            //   a no-no.
+            let error = parking_lot::Mutex::new(Some(error));
+            let reconnect = RetryIf::spawn(
+                ExponentialBackoff::from_millis(50),
+                || try_reconnect(inner.clone(), error.lock().clone()).map_err(Arc::new),
+                |err: &Arc<ConnectError>| {
+                    if let ConnectError::Canceled = **err {
+                        false
+                    } else {
+                        *error.lock() = Some(AcceptError::Reconnect(err.clone()));
+                        true
+                    }
+                },
+            );
             incoming = match reconnect.await {
                 Ok(incoming) => incoming,
                 Err(error) => {
                     debug!(%error, "reconnect failed, giving up");
-                    break error;
+                    break AcceptError::Reconnect(error);
                 }
             };
         }
     };
     for (_id, tun) in inner.load().tunnels.write().await.drain() {
-        let _ = tun.tx.send(Err(error)).await;
+        let _ = tun.tx.send(Err(error.clone())).await;
     }
 }